run_make_support: move target checks into `targets` module
